third_party/WebKit/LayoutTests/platform/win/virtual/mojo-loading/http/tests/security/contentSecurityPolicy/script-src-wildcards-disallowed-expected.txt

CONSOLE ERROR: line 21: Refused to load the script 'data:application/javascript,' because it violates the following Content Security Policy directive: "script-src 'nonce-nonce' *".

CONSOLE ERROR: line 36: Refused to load the script 'blob:http://127.0.0.1:8000/40814563-dbf6-4a8a-ad1c-de56e080f210' because it violates the following Content Security Policy directive: "script-src 'nonce-nonce' *".

CONSOLE ERROR: line 54: Refused to load the script 'filesystem:http://127.0.0.1:8000/temporary/fail.js' because it violates the following Content Security Policy directive: "script-src 'nonce-nonce' *".

This is a testharness.js-based test.
PASS data: URIs should not match * 
PASS blob: URIs should not match * 
PASS filesystem URIs should not match * 
Harness: the test ran to completion.
